Move GameState class to dedicated source file for more logical grouping
